Content length and cleanup
Added content length to the downloads. As part of this we now store the
ZIP file in a buffer; the upload is limited to 10MB in size so shouldn't
cause a memory problem. Eventually it should probably stream that out to
disk.

Also uploads will now be cleaned up after they have been used.